Switch to StatefulSet instead of Pod.
