+++ operator/src/main/kotlin/org/acasictf/ctf/operator/k8s.kt
@@ -1,6 +1,8 @@
 package org.acasictf.ctf.operator
 
 import io.fabric8.kubernetes.api.model.*
+import io.fabric8.kubernetes.api.model.apps.StatefulSet
+import io.fabric8.kubernetes.api.model.apps.StatefulSetSpec
 import io.fabric8.kubernetes.api.model.networking.v1.*
 
 fun meta(f: ObjectMeta.() -> Unit) = ObjectMeta().apply(f)
@@ -11,6 +13,10 @@ fun container(f: Container.() -> Unit) = Container().apply(f)
 
 fun pod(f: Pod.() -> Unit) = Pod().apply(f)
 fun podSpec(f: PodSpec.() -> Unit) = PodSpec().apply(f)
+fun podTemplateSpec(f: PodTemplateSpec.() -> Unit) = PodTemplateSpec().apply(f)
+
+fun statefulSet(f: StatefulSet.() -> Unit) = StatefulSet().apply(f)
+fun statefulSetSpec(f: StatefulSetSpec.() -> Unit) = StatefulSetSpec().apply(f)
 
 fun service(f: Service.() -> Unit) = Service().apply(f)
 fun serviceSpec(f: ServiceSpec.() -> Unit) = ServiceSpec().apply(f)
