+++ operator/src/main/kotlin/org/acasictf/ctf/operator/provisioner/kubernetes/creator/EnvCreator.kt
@@ -5,10 +5,10 @@ import org.acasictf.ctf.operator.model.kubernetes.v1alpha1.EnvTemplate
 import org.acasictf.ctf.operator.model.kubernetes.v1alpha1.Environment
 
 class EnvCreator(env: Environment, envTemplate: EnvTemplate, client: KubernetesClient) : Creator {
-    private val podCreator = PodCreator(env, envTemplate, client.pods())
+    private val statefulSetCreator = StatefulSetCreator(env, envTemplate, client.apps().statefulSets())
     private val serviceCreator = ServiceCreator(env, envTemplate, client.services())
     private val ingressCreator = IngressCreator(env, envTemplate, client.network().v1().ingresses())
-    private val creators = listOf(podCreator, serviceCreator, ingressCreator)
+    private val creators = listOf(statefulSetCreator, serviceCreator, ingressCreator)
 
     override fun create(dryRun: Boolean) {
         creators.forEach {
